max-bot/requirements.txt
# MAX Bot Dependencies
# Установка зависимостей только для MAX бота (микросервисная архитектура)

# MAX Bot Library
umaxbot>=0.1.0

# Configuration
python-dotenv>=1.0.0
pydantic>=2.5.0

# Logging and utilities
requests>=2.31.0

# Для интеграции с QA сервисом
grpcio>=1.60.0
grpcio-tools>=1.60.0


